requirements.txt
anime-face-detector>=0.0.4
mmcv-full==1.3.16
mmdet==2.18.0
mmpose==0.20.0
numpy>=1.21.3
opencv-python-headless>=4.5.4.58
openmim>=0.1.5
torch>=1.10.0
torchvision>=0.11.1
